remove old blender dependency
