+++ importing/importbuttons.py
@@ -11,8 +11,7 @@ This file performs the following operations
 .   Invokes the other import operations based on what options were chosen on the panel
 '''
 
-import bpy, os, time, glob, datetime
-from subprocess import Popen, PIPE
+import bpy, os, datetime
 
 from ..interface.dictionary_en import t
 from .. import common as c
@@ -103,9 +102,6 @@ class kkbp_import(bpy.types.Operator):
 
         #run functions based on selection
         c.toggle_console()
-        self.convert_and_import_textures()
-        c.toggle_console()
-        c.toggle_console() #have to toggle it twice after running the second blender instance
         self.import_pmx_models()
         for index, function in enumerate(functions):
             print('Import function {} running'.format(index))
@@ -145,21 +141,3 @@ class kkbp_import(bpy.types.Operator):
                             bpy.data.texts.remove(bpy.data.texts['Model_e'])
         c.initialize_timer()
         c.print_timer('Import PMX')
-    
-    def convert_and_import_textures(self):
-        c.kklog('Opening older version of Blender to convert model textures...')
-        time.sleep(5)
-        # You have to supply a blend file or it won't execute the script automatically. Choose the video editing template blend because it's the first one I tried
-        if 'blender.exe' in bpy.context.scene.kkbp.blender_path:
-            version_path = [i for i in glob.glob(os.path.dirname(bpy.context.scene.kkbp.blender_path) + '/*/')][0]
-        else:
-            bpy.context.scene.kkbp.blender_path = bpy.context.scene.kkbp.blender_path + '/blender.exe'
-            version_path = [i for i in glob.glob(os.path.dirname(bpy.context.scene.kkbp.blender_path) + '/*/')][0]
-        blender_file = os.path.join(version_path, 'scripts', 'startup', 'bl_app_templates_system', 'Video_Editing', 'startup.blend')
-        secondscriptname = os.path.join(os.path.dirname(__file__), 'converttextures.py')
-        process = Popen([bpy.context.scene.kkbp.blender_path, blender_file, "-P", secondscriptname, os.path.dirname(__file__), bpy.context.scene.kkbp.import_dir, '1'], stdout=PIPE, universal_newlines=True)
-        r = process.stdout.readline()[:-1]
-        while r:
-            if '|' in r:
-                c.kklog(r.replace('|','')) # these are lines printed from the second script
-            r = process.stdout.readline()[:-1]
